refactor : legacy 코드 제거
