+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -2,14 +2,19 @@ package com.techcourse.controller;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
+import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
+import web.org.springframework.web.bind.annotation.RequestMapping;
+import web.org.springframework.web.bind.annotation.RequestMethod;
+import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.view.JspView;
 
-public class RegisterController implements Controller {
+@Controller
+public class RegisterController {
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public ModelAndView save(HttpServletRequest req, HttpServletResponse res) {
         User user = new User(
                 2,
                 req.getParameter("account"),
@@ -18,7 +23,17 @@ public class RegisterController implements Controller {
         );
 
         InMemoryUserRepository.save(user);
-        return "redirect:/index.jsp";
+
+        return new ModelAndView(
+                new JspView("redirect:/index.jsp")
+        );
+    }
+
+    @RequestMapping(value = "/register", method = RequestMethod.GET)
+    public ModelAndView show(HttpServletRequest req, HttpServletResponse res) {
+        return new ModelAndView(
+                new JspView("/register.jsp")
+        );
     }
 
 }
