+++ app/src/main/java/com/techcourse/controller/UserController.java
@@ -29,4 +29,5 @@ public class UserController {
         modelAndView.addObject("user", user);
         return modelAndView;
     }
+
 }
