--- mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/asis/Controller.java
@@ -1,10 +1,0 @@
-package webmvc.org.springframework.web.servlet.mvc.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public interface Controller {
-
-    String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception;
-
-}
